adjusting routes & api files, order->sequence
