+++ backend/src/provision/dto/create-provision.dto.ts
@@ -11,6 +11,6 @@ export class CreateProvisionDto extends PickType(ProvisionDto, [
   'free_text',
   'help_text',
   'category',
-  'order_value',
+  'sequence_value',
   'create_userid',
 ] as const) {}
