+++ backend/src/provision/dto/update-provision.dto.ts
@@ -11,6 +11,6 @@ export class UpdateProvisionDto extends PickType(ProvisionDto, [
   'free_text',
   'help_text',
   'category',
-  'order_value',
+  'sequence_value',
   'update_userid',
 ] as const) {}
